Add root zone setup for scenario 3 (section 5.3 in RSSAC028)
This setup can be provisioned on servers-vm with the ansible playbook ansible/scenario-3_network_provision_servers-vm.yml

In this scenario, the root server addresses are added as direct A and AAAA records in the root zone, and signed as such. There is no delegation to a 'root server zone'. The particular setup chosen here is to also not use an ENT, i.e. the a root server's FQDN is 'a.', etc.