Add root zone setup for scenario 4 (section 5.4 in RSSAC028)
This setup can be provisioned on servers-vm with the ansible playbook ansible/scenario-4_network_provision_servers-vm.yml

In this scenario, there is no delegation through an externally hosted zone. The root servers have their addresses in a TLD specifically for this purpose.

In this particular setup, the TLD is 'some-servers', with the FQDN of the first root server being 'a.some-servers.'.