Add root zone setup for scenario 5 (section 5.5 in RSSAC028)
This setup can be provisioned on servers-vm with the ansible playbook ansible/scenario-5_network_provision_servers-vm.yml

In this particular setup, the added TLDs are 'a.' up to 'm.', signed and delegated from the root zone. The nameserver addresses within these zones are 'ns.<zone>. A' and 'ns.<zone>. AAAA'.